Extracted anthology expansion to own class: ExpandedParentPublication.java
